fix windows build from recent changes.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@3091 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/endpoints/mod_sofia/mod_sofia.c
@@ -870,7 +870,7 @@ char *encode_name(char *s)
 		*resource++ = '\0';
 	}
 
-	len = strlen(user) + 25;
+	len = (uint32_t)strlen(user) + 25;
 
 	if (at) {
 		*at++ = '\0';
